[FE] REFACT: AGUCabinetReturnSection container와 presenter로 분리 #1737
