+++ frontend/src/Cabinet/pages/ProfilePage.tsx
@@ -20,7 +20,6 @@ import {
 } from "@/Cabinet/api/axios/axios.custom";
 import instance from "../api/axios/axios.instance";
 
-
 const axiosLinkGoogleAccount = async (oauthMail: string) => {
   try {
     const { data } = await instance.post(
@@ -84,9 +83,9 @@ const ProfilePage = () => {
           <DisplayStyleCardContainer />
           <PointColorCardContainer />
           <AlarmCardContainer alarm={myInfo.alarmTypes} />
-          <SnsConnectionCardContainer
+          {/* <SnsConnectionCardContainer
             userOauthConnection={myInfo.userOauthConnection}
-          />
+          /> */}
         </CardGridWrapper>
       )}
     </>
